requirements.txt
Flask==3.0.0
peewee==3.16.3
psycopg2-binary==2.9.7
requests==2.31.0
urllib3==2.0.7
python-dotenv==1.0.0
gunicorn==21.2.0